Update bbc.js
BBC Weather Driver JSON API

*Script fetches data in JSON format;
*Script returns 7 days of forecast (actually it returns 14, but the desklet has the limit set at 7, the current day + 6 days of forecast);
*Script fetches data from 3 different APIs: Location, Observations and Forecasts;
*Allows insertion of geonameID(code city) and geographic coordinates (example: -4.05,39.66);
*Anticipated and synchronized fetching of APIs and parsing before assigning to objects;
*Delay in cleaning declared objects...Now the script only cleans objects when the API data is already parsed and ready for replacement, reducing the delay in refreshing the data;
*Adjustment of icons according to the data extracted from the APIs with numbering and nomenclature;
*Adjustment of nomenclature according to the desklet translation;
ToDo: add in system translation terms used in visibility data (Excellent, Very Good, Good, Moderate, Poor)

Author's note:
Creating this script was a huge challenge, given the limitations imposed by the desklet itself. Wxbase.js and desklet.js are not designed to work with multiple APIs, interleaved data retrieval, and synchronicity for script execution. I tried my best to stay within the scope of wxbase to maintain a standard, but I confess that I wrote hundreds of scripts that, if they were not designed as they are, would not work.
It was necessary to break the sequence chain determined by wxbase in getWeather so that the script could gather data from the 3 APIs and only then correctly assign the data to the objects without altering wxbase or desklet.js. The use of async/await was also essential so that the chaining of functions would bring the data that was essential to other parts of the script, so that the geographic location could be implemented. It took many attempts, and I had been working on this script for some time, but I finally got it. It works and it works well.
As a bonus, I designed the BBC logo in Gimp to use in the desklet. I didn't use the new font (reich font) but I used the old one, Gill Sans, which is no longer the official BBC font.